Update xsysmon_aux_polled_example.c
The value is saturated and incorrect when stored in "int" variable.
When i change "VAuxData" to float ir seems so resolve the problem.

+++ XilinxProcessorIPLib/drivers/sysmon/examples/xsysmon_aux_polled_example.c
@@ -148,7 +148,7 @@ int SysMonAuxPolledExample(u16 SysMonDeviceId)
 {
 	int Status;
 	XSysMon_Config *ConfigPtr;
-	u16 VAuxData[4];
+	float VAuxData[4];
 	u16 VAuxRawData[4];
 	XSysMon *SysMonInstPtr = &SysMonInst;
 	u32 Index;
